Fixed issue where 'logged in user status' in the header bar will only update upon refreshing the page. Fixed via React Context in AuthContext.jsx under the new src/context dir, wrapped the app around it in main.jsx, and updated the pages requiring it (header, login and register pages).
