+++ frontend-service/src/main.jsx
@@ -1,13 +1,16 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
+import { AuthProvider } from './context/AuthContext';
 import Header from './components/Header';
 import App from './App';
-import './styles.css'
+import './styles.css';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <BrowserRouter>
-  <Header />
-    <App />
+    <AuthProvider>
+      <Header />
+      <App />
+    </AuthProvider>
   </BrowserRouter>
 );
